Move bottom tray to components folder
