+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/components/bottomTray.tsx
