+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v1/table/index.tsx
@@ -31,7 +31,7 @@ import { Resizable } from 're-resizable';
 /**
  * Internal dependencies.
  */
-import BottomTray from './bottomTray';
+import BottomTray from '../../components/bottomTray';
 
 interface AuctionTableProps {
   auctionEvents: singleAuctionEvent[];
